Refactor env-vars
